project/requirements.txt
pillow
django
djangorestframework